requirements.txt
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
jupyter>=1.0.0
langgraph>=0.2.0
typing-extensions>=4.9.0
streamlit>=1.35.0
chromadb>=0.5.0
